新增上传记录保存返回主键

+++ mldong-sys-parent/mldong-sys-service/src/main/java/com/mldong/modules/sys/service/SysUploadRecordService.java
@@ -65,5 +65,5 @@ public interface SysUploadRecordService{
 	 * @param uploadRecord
 	 * @return
 	 */
-	public Long saveReturnId(SysUploadRecord uploadRecord);
+	public int saveReturnId(SysUploadRecord uploadRecord);
 }
